+++ unittest/libmariadb/features-10_2.c
@@ -190,7 +190,7 @@ static int com_multi_ps2(MYSQL *mysql)
   bind[0].buffer= &intval;
   bind[1].buffer_type= MYSQL_TYPE_STRING;
   bind[1].buffer= (char *)varval;
-  bind[1].buffer_length= strlen(varval);
+  bind[1].buffer_length= (ulong)strlen(varval);
 
   stmt= mysql_stmt_init(mysql);
   mysql_stmt_attr_set(stmt, STMT_ATTR_PREBIND_PARAMS, &param_count);
@@ -284,7 +284,7 @@ static int execute_direct_example(MYSQL *mysql)
   bind[0].buffer= &intval;
   bind[1].buffer_type= MYSQL_TYPE_STRING;
   bind[1].buffer= (char *)strval;
-  bind[1].buffer_length= strlen(strval);
+  bind[1].buffer_length= (ulong)strlen(strval);
 
   /* set number of parameters */
   if (mysql_stmt_attr_set(stmt, STMT_ATTR_PREBIND_PARAMS, &param_count))
